refactor(optionsmenu): remove fragment import, rename OptionsMenuUtil
