+++ client/src/components/OptionsMenu/OptionsSideMenu.component.tsx
@@ -1,13 +1,6 @@
 import { Box, Flex, Text, useMultiStyleConfig } from '@chakra-ui/react'
 import * as FullStory from '@fullstory/browser'
-import {
-  Fragment,
-  LegacyRef,
-  useEffect,
-  useState,
-  ReactElement,
-  createRef,
-} from 'react'
+import { LegacyRef, useEffect, useState, ReactElement, createRef } from 'react'
 import { useQuery } from 'react-query'
 import { Link as RouterLink } from 'react-router-dom'
 import { useGoogleAnalytics } from '../../contexts/googleAnalytics'
@@ -23,7 +16,7 @@ import {
   isSpecified,
   getTopicsQuery,
 } from '../../util/urlparser'
-import { bySpecifiedOrder } from './OptionsMenuUtil.component'
+import { bySpecifiedOrder } from './util'
 
 const OptionsSideMenu = ({ agency }: { agency?: Agency }): ReactElement => {
   const [hasTopicsKey, setHasTopicsKey] = useState(false)
@@ -89,7 +82,7 @@ const OptionsSideMenu = ({ agency }: { agency?: Agency }): ReactElement => {
     </Box>
   )
 
-  return <Fragment>{sideMenu}</Fragment>
+  return <>{sideMenu}</>
 }
 
 export default OptionsSideMenu
